Moved create user commands to users folder
