+++ src/Pelican.Application/Users/Commands/CreateAdmin/CreateAdminCommand.cs
@@ -1,4 +1,4 @@
 ﻿using Pelican.Application.Abstractions.Messaging;
 
-namespace Pelican.Application.Authentication.CreateAdmin;
+namespace Pelican.Application.Users.Commands.CreateAdmin;
 public sealed record CreateAdminCommand(string Name, string Email, string Password) : ICommand;
