+++ src/Pelican.Application/Users/Commands/CreateStandardUser/CreateStandardUserCommand.cs
@@ -1,5 +1,5 @@
 ﻿using Pelican.Application.Abstractions.Messaging;
 
-namespace Pelican.Application.Authentication.CreateStandardUser;
+namespace Pelican.Application.Users.Commands.CreateStandardUser;
 public sealed record CreateStandardUserCommand(string Name, string Email, string Password) : ICommand;
